// packages/40009/src copy/apis/register.ts
export interface MedicalPayType {
  ordStas: string;
  hisName: string;
  psnAcctPay: number;
  feeSumamt: number;
  subHisName: string;
  extData: {
    preSetl: Partial<{
      cvlserv_pay: number;
      cvlserv_flag: string;
      med_type: string;
      brdy: string;
      naty: string;
      psn_cash_pay: number;
      certno: string;
      hifmi_pay: number;
      psn_no: string;
      act_pay_dedc: number;
      mdtrt_cert_type: string;
      balc: number;
      medins_setl_id: string;
      psn_cert_type: string;
      acct_mulaid_pay: number;
      clr_way: string;
      hifob_pay: number;
      oth_pay: number;
      medfee_sumamt: number;
      hifes_pay: number;
      gend: string;
      mdtrt_id: string;
      acct_pay: number;
      fund_pay_sumamt: number;
      fulamt_ownpay_amt: number;
      hosp_part_amt: number;
      inscp_scp_amt: number;
      insutype: string;
      maf_pay: number;
      psn_name: string;
      psn_part_amt: number;
      pool_prop_selfpay: number;
      psn_type: string;
      hifp_pay: number;
      overlmt_selfpay: number;
      preselfpay_amt: number;
      age: number;
      clr_type: string;
    }>;
  };
  othFeeAmt: number;
  deposit: number;
  fundPay: number;
  medicalReq: Partial<{
    acctUsedFlag: string;
    atddrNo: string;
    backUrl: string;
    begntime: string;
    caty: string;
    chrgBchno: string;
    deptCode: string;
    deptName: string;
    diseCodg: string;
    diseName: string;
    diseinfoList: string;
    drName: string;
    ecToken: string;
    feeType: string;
    feedetailList: string;
    fetts: number;
    fetusCnt: number;
    fulamtOwnpayAmt: string;
    gesoVal: number;
    idNo: string;
    idType: string;
    insuCode: string;
    iptOtpNo: string;
    mainCondDscr: string;
    matnType: string;
    mdtrtCertType: string;
    medOrgOrd: string;
    medType: string;
    medfeeSumamt: string;
    medicalCardId: string;
    medicalCardInstId: string;
    obid: string;
    orgCodg: string;
    payAuthNo: string;
    psnSetlway: string;
    rxCircFlag: string;
    spbillCreateIp: string;
    uldLatlnt: string;
    userName: string;
  }>;
  payOrdId: string;
  ownPayAmt: number;
}